refactor: separate auth expiry notifications
